backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email]
python-jose[cryptography]
bcrypt
python-multipart
psycopg2-binary        # For PostgreSQL
gunicorn   
